Поправила тесты на вход в систему
